+++ router/floors_route.js
@@ -1,5 +1,6 @@
 module.exports = function(app) {
 	var mongoose = require('mongoose'),
+		adminhash = require('./pwd.js'),
 		Floors = mongoose.model('floor', {
 			coll: {
 				type: String,
@@ -62,14 +63,26 @@ module.exports = function(app) {
 			res.json(floors);
 		});
 	}).post((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		Floors.create(req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).put((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		Floors.findByIdAndUpdate(req.params.id, req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).delete((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		Floors.findByIdAndRemove(req.params.id, (err, floor) => {
 			res.json(floor);
 		});
@@ -85,14 +98,26 @@ module.exports = function(app) {
 			res.json(floors);
 		});
 	}).post((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorColls.create(req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).put((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorColls.findByIdAndUpdate(req.params.id, req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).delete((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorColls.findByIdAndRemove(req.params.id, (err, floor) => {
 			res.json(floor);
 		});
@@ -108,14 +133,26 @@ module.exports = function(app) {
 			res.json(floors);
 		});
 	}).post((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorModels.create(req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).put((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorModels.findByIdAndUpdate(req.params.id, req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).delete((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorModels.findByIdAndRemove(req.params.id, (err, floor) => {
 			res.json(floor);
 		});
@@ -131,14 +168,26 @@ module.exports = function(app) {
 			res.json(floors);
 		});
 	}).post((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorGammas.create(req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).put((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorGammas.findByIdAndUpdate(req.params.id, req.body, (err, floor) => {
 			res.json(floor);
 		});
 	}).delete((req, res) => {
+		if (adminhash !== req.cookies.adminhash) {
+			res.status(403).send('Not authorized');
+			return;
+		}
 		FloorGammas.findByIdAndRemove(req.params.id, (err, floor) => {
 			res.json(floor);
 		});
